requirements.txt
matplotlib==3.1.3
numpy==1.18.1
opencv_python==4.5.5.64
Pillow==9.4.0
PyYAML==6.0
